requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cloudinary==1.44.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
